requirements.txt
keras==2.8.0
Keras-Preprocessing==1.1.2
matplotlib==3.5.2
matplotlib-inline==0.1.2
numpy==1.22.3
opencv-python==4.5.5.64
pandas==1.4.2
Pillow==9.1.0
scikit-learn==1.0.2
scipy==1.8.0
seaborn==0.11.2
tensorboard==2.8.0
tensorboard-plugin-wit==1.8.1
streamlit==1.9.0
tensorflow==2.8.0
